changed page request to POST method & get subcategories by category & added "done" attribute to order search request & added username to order

+++ src/main/java/roman/pidkostelnyi/victoriaarmario/tool/FileTool.java
@@ -31,6 +31,9 @@ public class FileTool {
     }
 
     public boolean deleteFile(String directory, String fileName) {
+        if (fileName == null || fileName.isEmpty()) {
+            return true;
+        }
         return Paths.get(System.getProperty(USER_HOME), directory, fileName).toFile().delete();
     }
